Add filter-pcb for denoising stepper signals
When the VFD was running the stepper driver did not work correctly.
The especially the DIR signal got unusable and triggered randomly.

Add custom stripboard that reduces the noise on DIR and PULSE data
lines.
The circuit simply increases the current in the data lines significantly
and crops the voltage to 0-5V using a zener diode.
While testing with an oscilloscope the circuit seems effective.

- add kicad schematic
- add spreadsheet with stripboard layout
- update connection: plan with stripboard, update changed 5V connection